Move build-message rendering into log.ts
`BuildError.ts` had grown to mix two responsibilities: the error data
model (the `BuildError` union and the `BundlerError` adapter) and the
console rendering of those errors. Adding warning rendering for
unresolved workspace dependencies made the split worse — the module was
no longer just about errors, yet the rendering didn't belong with the
error types either.

Move all the rendering and logging into `log.ts`, which already owns the
CLI's styled console output (path styling, file/function/status logs).
The build-message renderers are the same kind of code and sit naturally
alongside the existing `logFailure` / `logWarn` helpers. `BuildError.ts`
keeps only the error model and the bundler adapter.

The dependency runs one way — `log.ts` imports the error types from
`BuildError.ts` (type-only), and the model no longer pulls in `log.ts`
for `formatPathDoc` — so no import cycle.
